* Set $ENV{AUTOMATED_TESTING} with -a, skip prereqs with -p, skip kwalitee with -k
H. Merijn Brand wanted to set AUTOMATED_TESTING, so I added the -a switch
and the automated_testing configuration file directive to do that.

And, you can skip prereq tests with -p on the command line or skip_prereqs in
the config file. Similarly, -k and skip_kwalitee. That used to be just 'kwalitee'
but I made it parallel to skip_prereqs. You have to turn things off, so
the directives say that.