Modify Module::Release->dist_version to parse 3-part version strings
Currently if a version like v0.1.0 as generated by Module::Build if
PBP is followed and a version like $VERSION = qv('0.1') is specified,
this is mis-parsed as "1.0".

Try to fix this when 3 or or parts are detected, or when a 'v' prefix
is seen, by parsing it using version.pm if available, or using the
version verbatim if not.