Require App::CPANTS::Lint for kwalitee checks
The cpants_lint.pl program has been extracted into a separate project
and hence one needs `App::CPANTS::Lint` in order that the kwalitee tests
in `Module::Release` can be run.  Here I've extracted the required
module names into variables to try and be more descriptive, but also to
reduce the length of the strings required in the `eval` and in the `die`
message.