Replace plain terminal input with new _slurp() method
This change simply adds infrastructure to allow for the `<>` input to be
more easily overridden in later tests by overriding the new `_slurp`
method.  It seems that one would have to jump through many hoops in
order to make `<>` overridable so that one could test its behaviour and
that of e.g. `get_env_var`, hence introducing a new method seemed the
best option.  The new method is used in `get_env_var` to replace direct
usage of the diamond operator.  This change also adds a simple test of
the new method which checks and documents its behaviour.