Avoid echoing passwords to the terminal
If we want to enter the CPAN_PASS value, we don't want to see it echoed
to the terminal.  The implementation used here is based off that used
for the `cpan-upload` script from `CPAN::Uploader`.  Unfortunately,
testing that the value isn't echoed is very difficult, hence the tests
added here only exercise the code path using the CPAN_PASS variable and
check that the value returned from `get_env_var` is still that that one
would expect.