Return release date string via get_release_date() method
The release date given in the Changes file should conform to the
format defined in CPAN::Changes::Spec.  Although the code in the
`release` script already conformed to this spec, it's nicer to have the
functionality wrapped in a method.  These changes add the new
`get_release_date()` method as well as tests for it, and replace the
existing code in the `release` script with the new method call.