Show list of contributors since previous release
These changes implement two routines which fulfill the desired feature
mentioned in #11: one to print the list of recent contributors
(`show_recent_contributors()`) and one to obtain the list of recent
contributors from the relevant version control system
(`get_recent_contributors()`).  The `get_recent_contributors()` method
is only a stub and needs to be overridden by the appropriate mixin for
the given version control system.  A concrete implementation of this
method will be submitted in a separate PR to the `Module::Release::Git`
repository.